Add omdb, change api, add layers, add linter, add pre-commit

+++ movie/views.py
@@ -1,15 +1,13 @@
-from dataclasses import asdict
-
 from django.db import IntegrityError
-from django.db.models import Count, OuterRef, Exists
+from django.db.models import Count, Exists, OuterRef
 from django_filters.rest_framework import DjangoFilterBackend
-from drf_spectacular.utils import extend_schema, OpenApiParameter
+from drf_spectacular.utils import extend_schema
 from rest_framework import permissions, status
 from rest_framework.filters import OrderingFilter
 from rest_framework.generics import (
-    GenericAPIView,
-    DestroyAPIView,
     CreateAPIView,
+    DestroyAPIView,
+    GenericAPIView,
     ListAPIView,
     RetrieveAPIView,
 )
@@ -17,39 +15,33 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from throttling.throttling import (
-    RegularSearchUaThrottle,
-    RegularSearchIpThrottle,
     RegularSearchForwardedThrottle,
-    AiSearchUaThrottle,
-    AiSearchIpThrottle,
-    AiSearchForwardedThrottle,
+    RegularSearchIpThrottle,
+    RegularSearchUaThrottle,
 )
-from .ai_find_movie import FindMovieAiClient
+
 from .errors import AddLikeError
 from .filters import MovieFilter, WatchLaterFilter
-from .models import Movie, WatchLaterMovie, LikeMovie, Director, Actor, Genre
+from .models import LikeMovie, Movie, WatchLaterMovie
 from .paginations import MoviesPagination
 from .serializers import (
+    FindMovieAiSearchViewRequestSerializer,
+    FindMovieSearchViewRequestSerializer,
+    MovieModelSerializer,
     MovieSerializer,
     WatchLaterCreateSerializer,
     WatchLaterListSerializer,
-    FindMovieAiViewRequestSerializer,
 )
 from .services import MovieService
 
 
 class MovieView(RetrieveAPIView):
     permission_classes = [permissions.AllowAny]
-    serializer_class = MovieSerializer
-    lookup_field = "id"
+    serializer_class = MovieModelSerializer
+    lookup_field = 'id'
 
     def get_queryset(self):
-        return (
-            Movie.objects.with_is_liked(self.request.user.id)
-            .with_is_watch_later(self.request.user.id)
-            .with_likes_count()
-            .with_watch_later_count()
-        )
+        return Movie.objects.with_is_liked(self.request.user.id).with_is_watch_later(self.request.user.id).with_likes_count().with_watch_later_count()
 
 
 class MovieLikeView(CreateAPIView):
@@ -57,7 +49,7 @@ class MovieLikeView(CreateAPIView):
 
     def post(self, request, *args, **kwargs):
         try:
-            LikeMovie.objects.create(user=self.request.user, movie_id=kwargs.get("id"))
+            LikeMovie.objects.create(user=self.request.user, movie_id=kwargs.get('id'))
         except IntegrityError:
             raise AddLikeError
 
@@ -69,16 +61,14 @@ class MovieUnlikeView(GenericAPIView):
     queryset = LikeMovie.objects.all()
 
     def post(self, request, *args, **kwargs):
-        LikeMovie.objects.filter(
-            user=self.request.user, movie_id=kwargs.get("id")
-        ).delete()
+        LikeMovie.objects.filter(user=self.request.user, movie_id=kwargs.get('id')).delete()
 
         return Response(status=status.HTTP_200_OK)
 
 
-class FindMovieView(ListAPIView):
+class MoviesListView(ListAPIView):
     permission_classes = [permissions.AllowAny]
-    serializer_class = MovieSerializer
+    serializer_class = MovieModelSerializer
     pagination_class = MoviesPagination
     throttle_classes = [
         RegularSearchUaThrottle,
@@ -87,107 +77,61 @@ class FindMovieView(ListAPIView):
     ]
     filter_backends = [DjangoFilterBackend, OrderingFilter]
     filterset_class = MovieFilter
-    search_fields = ["title"]
-    ordering_fields = ["imdb_id", "title", "genre", "year", "likes_count"]
-    ordering = ["imdb_id"]
+    search_fields = ['title']
+    ordering_fields = ['imdb_id', 'title', 'genre', 'year', 'likes_count']
+    ordering = ['imdb_id']
 
     def get_queryset(self):
-        queryset = Movie.objects.all()
-        if expression := self.request.query_params.get("expression"):
-            queryset = queryset.filter(title__icontains=expression)
         return (
-            queryset.with_is_liked(self.request.user.id)
+            Movie.objects.all()
+            .with_is_liked(self.request.user.id)
             .with_is_watch_later(self.request.user.id)
             .with_likes_count()
             .with_watch_later_count()
         )
 
-    @extend_schema(parameters=[OpenApiParameter(name="expression", type=str, location=OpenApiParameter.QUERY)])
-    def get(self, *args, **kwargs):
-        if expression := self.request.query_params.get("expression"):
-            movies = MovieService.get_movies_from_imdb(expression)
-            Movie.objects.bulk_create(
-                [Movie(**asdict(movie)) for movie in movies],
-                update_conflicts=True,
-                unique_fields=["title"],
-                update_fields=["imdb_id", "poster", "year", "type"],
-            )
-        return super().get(*args, **kwargs)
 
-
-class FindMovieAiView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
-    serializer_class = MovieSerializer
+class MoviesSearchView(APIView):
+    permission_classes = [permissions.AllowAny]
     throttle_classes = [
-        AiSearchUaThrottle,
-        AiSearchIpThrottle,
-        AiSearchForwardedThrottle,
+        RegularSearchUaThrottle,
+        RegularSearchIpThrottle,
+        RegularSearchForwardedThrottle,
     ]
 
     @extend_schema(
-        request=FindMovieAiViewRequestSerializer,
+        request=FindMovieSearchViewRequestSerializer,
         responses={status.HTTP_200_OK: MovieSerializer(many=True)},
     )
     def post(self, request, *args, **kwargs):
-        # TODO: Move this to a service
-        input_serializer = FindMovieAiViewRequestSerializer(data=request.data)
+        input_serializer = FindMovieSearchViewRequestSerializer(data=request.data)
         input_serializer.is_valid(raise_exception=True)
-        ai_movies = FindMovieAiClient(**input_serializer.data).find_movies()
-        for ai_movie in ai_movies:
-            imdb_movie = next(
-                filter(
-                    lambda x: x.title == ai_movie.title,
-                    MovieService.get_movies_from_imdb(ai_movie.title),
-                ),
-                None,
-            )
-            ai_movie.poster = imdb_movie.poster
+        movie_service = MovieService()
+        imdb_movies = movie_service.get_movies_from_imdb(input_serializer.data.get('expression'))
+        omdb_movies = movie_service.search_movies_in_omdb([movie.title for movie in imdb_movies], self.request.user.id)
+        serialized_movies = MovieSerializer(omdb_movies, many=True)
+        return Response(serialized_movies.data, status=status.HTTP_200_OK)
 
-        director_names = {m.director for m in ai_movies}
-        genre_names = {g for m in ai_movies for g in m.genre}
-        actor_names = {a for m in ai_movies for a in m.actors}
 
-        existing_directors = {d.full_name: d for d in Director.objects.filter(full_name__in=director_names)}
-        existing_genres = {g.name: g for g in Genre.objects.filter(name__in=genre_names)}
-        existing_actors = {a.full_name: a for a in Actor.objects.filter(full_name__in=actor_names)}
+class MoviesAiSearchView(APIView):
+    permission_classes = [permissions.IsAuthenticated]
+    # throttle_classes = [
+    #     AiSearchUaThrottle,
+    #     AiSearchIpThrottle,
+    #     AiSearchForwardedThrottle,
+    # ]
 
-        new_directors = [Director(full_name=name) for name in director_names if name not in existing_directors]
-        new_genres = [Genre(name=name) for name in genre_names if name not in existing_genres]
-        new_actors = [Actor(full_name=name) for name in actor_names if name not in existing_actors]
-
-        Director.objects.bulk_create(new_directors)
-        Genre.objects.bulk_create(new_genres)
-        Actor.objects.bulk_create(new_actors)
-
-        existing_directors.update({d.full_name: d for d in Director.objects.filter(full_name__in=director_names)})
-        existing_genres.update({g.name: g for g in Genre.objects.filter(name__in=genre_names)})
-        existing_actors.update({a.full_name: a for a in Actor.objects.filter(full_name__in=actor_names)})
-
-        movies = []
-        for ai_movie in ai_movies:
-            movie = Movie(
-                title=ai_movie.title,
-                imdb_id=ai_movie.imdb_id,
-                poster=ai_movie.poster,
-                year=ai_movie.year,
-                type=ai_movie.type,
-                plot=ai_movie.plot,
-                director=existing_directors[ai_movie.director],
-            )
-            movies.append((movie, ai_movie.genre, ai_movie.actors))
-
-        Movie.objects.bulk_create(
-            [m for m, _, _ in movies],
-            update_conflicts=True,
-            unique_fields=["title"],
-            update_fields=["imdb_id", "poster", "year", "type", "plot"],
-        )
-
-        for movie, genres, actors in movies:
-            movie.genres.set([existing_genres[g] for g in genres])
-            movie.actors.set([existing_actors[a] for a in actors])
-
-        serialized_movies = MovieSerializer([m for m, _, _ in movies], many=True)
+    @extend_schema(
+        request=FindMovieAiSearchViewRequestSerializer,
+        responses={status.HTTP_200_OK: MovieSerializer(many=True)},
+    )
+    def post(self, request, *args, **kwargs):
+        input_serializer = FindMovieAiSearchViewRequestSerializer(data=request.data)
+        input_serializer.is_valid(raise_exception=True)
+        movie_service = MovieService()
+        ai_movies = movie_service.get_movies_from_ai(input_serializer.data.get('expression'))
+        omdb_movies = movie_service.search_movies_in_omdb([movie.title for movie in ai_movies], self.request.user.id)
+        serialized_movies = MovieSerializer(omdb_movies, many=True)
         return Response(serialized_movies.data, status=status.HTTP_200_OK)
 
 
@@ -197,7 +141,7 @@ class WatchLaterCreateView(CreateAPIView):
 
     def get_serializer_context(self):
         context = super().get_serializer_context()
-        context["user"] = self.request.user
+        context['user'] = self.request.user
         return context
 
 
@@ -207,26 +151,22 @@ class WatchLaterListView(ListAPIView):
     pagination_class = MoviesPagination
     filter_backends = [DjangoFilterBackend, OrderingFilter]
     filterset_class = WatchLaterFilter
-    search_fields = ["movie__title"]
-    ordering_fields = ["movie__imdb_id", "movie__title", "movie__genre", "movie__year"]
-    ordering = ["movie__imdb_id"]
+    search_fields = ['movie__title']
+    ordering_fields = ['movie__imdb_id', 'movie__title', 'movie__genre', 'movie__year']
+    ordering = ['movie__imdb_id']
 
     def get_queryset(self):
         return (
             WatchLaterMovie.objects.filter(user_id=self.request.user.id)
-            .select_related("movie")
+            .select_related('movie')
             .annotate(
-                likes_count=Count("movie__likemovie"),
+                likes_count=Count('movie__likemovie'),
                 is_liked=Exists(
-                    LikeMovie.objects.filter(
-                        user_id=self.request.user.id, movie_id=OuterRef("movie__pk")
-                    ),
+                    LikeMovie.objects.filter(user_id=self.request.user.id, movie_id=OuterRef('movie__pk')),
                 ),
-                watch_later_count=Count("movie__watchlatermovie"),
+                watch_later_count=Count('movie__watchlatermovie'),
                 is_watch_later=Exists(
-                    WatchLaterMovie.objects.filter(
-                        user_id=self.request.user.id, movie_id=OuterRef("movie__pk")
-                    ),
+                    WatchLaterMovie.objects.filter(user_id=self.request.user.id, movie_id=OuterRef('movie__pk')),
                 ),
             )
         )
@@ -236,8 +176,6 @@ class WatchLaterDestroyView(DestroyAPIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def delete(self, request, *args, **kwargs):
-        WatchLaterMovie.objects.filter(
-            user_id=self.request.user.id, movie_id=kwargs.get("pk")
-        ).delete()
+        WatchLaterMovie.objects.filter(user_id=self.request.user.id, movie_id=kwargs.get('pk')).delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
